backend/requirements.txt
fastapi
uvicorn
deepface
opencv-python
numpy
bcrypt
python-dotenv
sqlalchemy
python-jose
python-multipart
cryptography
